Испавление сообщений ошибок

+++ app/Http/Controllers/ServicesController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers;
 use App\Service;
 use App\User;
 use App\Service_type;
-use App\User_service;
 use App\Feedbacks;
 use App\Favorites;
 use Illuminate\Http\Request;
@@ -21,8 +20,12 @@ class ServicesController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+     // Attēlot pakalpojumu lapu
     public function index()
     {
+        // Izveidot jaunu pieprasījumu datu bāzei
+        // Un pievienot filtrus, kas ir meklēšanas ievadlauku vērtības
+        // Katrs lauks tiek pārbaudīts uz tukšumu
         $services = Service::query();
         if (Input::has('title') && !empty(Input::get('title'))) {
             $services->where('title', 'like', '%' . Input::get('title') . '%');
@@ -41,32 +44,51 @@ class ServicesController extends Controller
             $services->where('district', '=', Input::get('district'));
         }
 
-        $services = $services->orderBy('updated_at', 'desc')->paginate(10);
+        $services = $services->orderBy('title')->paginate(10);
         $serviceTypes = Service_type::orderBy('name')->get();
 
         foreach ($services as $service) {
+            // Attēlot pirmos 100 simbolus
             $service->description = str_limit($service->description, 100);
         }
 
         return view('Service/pakalpojumi')->with('services', $services)->with('serviceTypes', $serviceTypes);
     }
 
-    public function sort($id)
+    // Filtrēt pakalpojumus pēc to veida
+    public function filter($id)
     {
         $services = Service::query();
+        // Princips kā index() funkcijai, tikai tiek iegūti pakalpojumi
+        // ar pareizo pakalpojuma veidu
         $services->where('service_type_id', '=', $id);
-        if (Input::has('title')) {
-            echo 'Has title';
+
+        if (Input::has('title') && !empty(Input::get('title'))) {
             $services->where('title', 'like', '%' . Input::get('title') . '%');
         }
-        $services = $services->orderBy('updated_at', 'desc')->paginate(10);
+        if (Input::has('fromPrice') && !empty(Input::get('fromPrice'))) {
+            $services->where('price', '>=', Input::get('fromPrice'));
+        }
+        if (Input::has('toPrice') && !empty(Input::get('toPrice'))) {
+            $services->where('price', '<=', Input::get('toPrice'));
+        }
+        if (Input::has('peopleCount') && !empty(Input::get('peopleCount'))) {
+            $services->where('person_number_from', '<=', Input::get('peopleCount'))
+                        ->where('person_number_to', '>=', Input::get('peopleCount'));
+        }
+        if (Input::has('district') && Input::get('district') != "Visi") {
+            $services->where('district', '=', Input::get('district'));
+        }
+        $services = $services->orderBy('title')->paginate(10);
         $serviceTypes = Service_type::orderBy('name')->get();
         foreach ($services as $service) {
+            // Attēlot pirmos 100 simbolus
             $service->description = str_limit($service->description, 100);
         }
         return view('Service/pakalpojumi')->with('services', $services)->with('serviceTypes', $serviceTypes);
     }
 
+    // Attēlot  favorītus
     public function favorites()
     {
         $userFavorites = Favorites::where('user_id', '=', Auth::id())->paginate(10);
@@ -78,6 +100,7 @@ class ServicesController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+    // Pakalpojuma izveidošanas skats
     public function create()
     {
         $serviceTypes = Service_type::orderBy('name')->pluck('name', 'id');
@@ -90,24 +113,27 @@ class ServicesController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
+    // Izveidot pakalpojumu
     public function store(Request $request)
     {
+        // Lauku validācijas nosacījumi
         $rules = array(
-            'title' => 'required|string|max:100',
-            'extra_service' => 'string|max:1000',
-            'district' => 'required|string|max:50',
-            'adress' => 'string|max:100',
-            'price' => 'required|numeric|min:0',
-            'email' => 'required|string|max:256',
-            'phone' => 'required|string|max:20',
-            'person_number_from' => 'digits_between:1,100000',
-            'person_number_to' => 'integer|min:' . $request->person_number_from,
+            'title' => 'required|string|between:1,100',
+            'extra_service' => 'nullable|max:1000',
+            'district' => 'required|string',
+            'adress' => 'nullable|max:100',
+            'price' => 'required|numeric|between:0.01,99999999999999999999',
+            'email' => 'required|string|between:6,256',
+            'phone' => 'required|string|between:6,20',
+            'person_number_from' => 'nullable|integer',
+            'person_number_to' => 'nullable|integer|min:' . $request->person_number_from,
             'description' => 'required|string|max:5000',
             'photo' => 'required|image',
         );
 
         $this->validate($request, $rules);
 
+        // Saglabāt failu
         $fileNameWithExt = $request->file('photo')->GetClientOriginalName();
         $filename = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
         $extension = $request->file('photo')->getClientOriginalExtension();
@@ -122,14 +148,22 @@ class ServicesController extends Controller
         $service->price = $request->input('price');
         $service->email = $request->input('email');
         $service->phone = $request->input('phone');
-        $service->person_number_from = $request->input('person_number_from');
-        $service->person_number_to = $request->input('person_number_to');
+        if($request->input('person_number_from') == null){
+            $service->person_number_from = 0;
+        } else {
+            $service->person_number_from = $request->input('person_number_from');
+        }
+        if($request->input('person_number_to') == null){
+            $service->person_number_to = 0;
+        } else {
+            $service->person_number_to = $request->input('person_number_to');
+        }
         $service->description = $request->input('description');
         $service->photo = $fileNameToStore;        
         $service->service_type_id = $request->input('serviceType');        
         $service->save();
 
-        return redirect()->action('ServicesController@index')->with('success', 'Pakalpoums pievienots');
+        return redirect()->action('ServicesController@index')->with('success', 'Pakalpojums pievienots');
     }
 
     /**
@@ -138,6 +172,7 @@ class ServicesController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+    // Attēlot pakalpojumu
     public function show($id)
     {
         $service = Service::find($id);
@@ -157,6 +192,7 @@ class ServicesController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+    // Refiģēt pakalpojumu skats
     public function edit($id)
     {
         $service = Service::find($id);
@@ -173,18 +209,20 @@ class ServicesController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+    // Rediģēt pakalpojumu
     public function update(Request $request, $id)
     {
+        // Validācijas noteikumi
         $rules = array(
-            'title' => 'required|string|max:100',
-            'extra_service' => 'string|max:1000',
-            'district' => 'required|string|max:50',
-            'adress' => 'string|max:100',
-            'price' => 'required|numeric|min:0',
-            'email' => 'required|string|max:256',
-            'phone' => 'required|string|max:20',
-            'person_number_from' => 'digits_between:1,100000',
-            'person_number_to' => 'integer|min:' . $request->person_number_from,
+            'title' => 'required|string|between:1,100',
+            'extra_service' => 'nullable|max:1000',
+            'district' => 'required|string',
+            'adress' => 'nullable|max:100',
+            'price' => 'required|numeric|between:0.01,99999999999999999999',
+            'email' => 'required|string|between:6,256',
+            'phone' => 'required|string|between:8,20',
+            'person_number_from' => 'nullable|integer',
+            'person_number_to' => 'nullable|integer|min:' . $request->person_number_from,
             'description' => 'required|string|max:5000',
             'photo' => 'image',
         );
@@ -193,6 +231,7 @@ class ServicesController extends Controller
 
         $service = Service::find($id);
         
+        // Ja jauna bilde tika ievietota, tad dzēst veco un ielikt jauno
         if ($request->hasFile('photo')) {
             $fileNameWithExt = $request->file('photo')->GetClientOriginalName();
             $filename = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
@@ -210,9 +249,18 @@ class ServicesController extends Controller
         $service->price = $request->input('price');
         $service->email = $request->input('email');
         $service->phone = $request->input('phone');
-        $service->person_number_from = $request->input('person_number_from');
-        $service->person_number_to = $request->input('person_number_to');
+        if($request->input('person_number_from') == null){
+            $service->person_number_from = 0;
+        } else {
+            $service->person_number_from = $request->input('person_number_from');
+        }
+        if($request->input('person_number_to') == null){
+            $service->person_number_to = 0;
+        } else {
+            $service->person_number_to = $request->input('person_number_to');
+        }
         $service->description = $request->input('description');
+        $service->service_type_id = $request->input('serviceType');        
         $service->save();
 
         return redirect()->action('ServicesController@index')->with('success', 'Pakalpoums atjaunots');
@@ -224,6 +272,7 @@ class ServicesController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+    // Izdzēst pakalpojumu
     public function destroy($id)
     {
         $service = Service::find($id);
@@ -232,20 +281,40 @@ class ServicesController extends Controller
         return redirect('/services')->with('success', 'Pakalpojums dzēsts');
     }
 
+    // Atvērt rakstīšanas skatu e-ziņai, kura tiks aizsūtīta administratoriem
+    // par pakalpojumu, kuru vēlas piedāvāt pievienošanai
     public function request()
     {
         return view('E-mail/pakalpojums-admin');
     }
 
+    // Aizsūtīt ziņu administratoriem
     public function emailRequest(Request $request)
     {
+        // Validācijas noteikumi
+        $rules = array(
+            'content' => 'between:1,7000',
+            'photo' => 'required|image',
+        );
+
+        $this->validate($request, $rules);
+
+        // Iegūt ziņu un bildes nosaukumu
         $content = $request->input('content');
         $pictureName = $request->file('photo')->GetClientOriginalName();
+
+        // Iegūt visus administratorus
         $admins = User::where('role_id', '=', 1)->get();
-        
-        Mail::send('E-mail.requestService', ['mailer' => Auth::user()->email, 'content' => $content], function ($message) use ($request, $pictureName) {
+        $adminEmails = array();
+        //Ielikt administratoru e-pastus masīvā
+        foreach($admins as $admin){
+            array_push($adminEmails, $admin->email);
+        }
+
+        //Nosūtīt ziņu
+        Mail::send('E-mail.requestService', ['mailer' => Auth::user()->email, 'content' => $content], function ($message) use ($request, $pictureName, $adminEmails) {
             $message->from(Auth::user()->email, Auth::user()->name . ' ' . Auth::user()->surname);
-            $message->to($admins);
+            $message->to($adminEmails);
             $message->attach($request->photo, ['as' => $pictureName]);
             $message->subject("Bankett pakalpojuma pieprasījums");
         });
@@ -253,33 +322,50 @@ class ServicesController extends Controller
         return redirect()->action('ServicesController@index');
     }
 
+    // Attēlot skatu, lai sazināties ar pakalpojuma sniedzēju
     public function contact($id){
         return view('E-mail/e-zina')->with('serviceId', $id);
     }
 
+    // Nosūtīt ziņu pakalpojuma sniedzējam
     public function emailContact(Request $request)
     {
+        // Validācijas noteikumi
+        $rules = array(
+            'content' => 'between:1,5000',
+        );
+
+        $this->validate($request, $rules);
+
         $content = $request->input('content');
         $serviceId = $request->input('serviceId');
         $service = Service::find($serviceId);
         $contactEmail = $service->email;
         $serviceName = $service->title;
 
+        // Aizsūtīt ziņu
         Mail::send('E-mail.contactService', ['mailer' => Auth::user()->email, 'content' => $content, 'serviceName' => $serviceName], function ($message) use ($contactEmail) {
             $message->from(Auth::user()->email, Auth::user()->name . ' ' . Auth::user()->surname);
             $message->to($contactEmail);
             $message->subject("Bankett ziņa");
         });
 
-        return redirect()->action('ServicesController@index');
+        return redirect()->action('ServicesController@index')->with('success', 'Jūsu e-ziņa ir iesutīta pakalpojua sniedzējam');;
     }
 
+    // Pievienot atsauksmi
     public function storeFeedback(Request $request, $id){
+        // Attēlot kļūdu tekstu latviešu valodā
+        $messages = [
+            'between' => 'Ievadītai atsauksmei jābūt no :min līdz :max simbolu skaitam',
+        ];
+
+        // Validācijas noteikumi
         $rules = array(
-            'text' => 'required|string|max:1000',
+            'text' => 'between:1,1000',
         );
 
-        $this->validate($request, $rules);
+        $this->validate($request, $rules, $messages);
 
         $feedback = new Feedbacks;
         $feedback->text = $request->input('text');
@@ -290,29 +376,40 @@ class ServicesController extends Controller
         return redirect()->action('ServicesController@show', ['id' => $id])->with('success', 'Atsauksme pievienota');
     }
 
+    // Izdzēst atsauksmi
     public function deleteFeedback($id){
         $feedback = Feedbacks::find($id);
         $serviceId = $feedback->service_id;
+
+        // Ja lietotājs nav administrators vai atsauksmes veidotājs
+        // Tad attēlot kļūdu
+        if(!(Auth::user()->isAdmin()) && Auth::user()->id != $feedback->user_id){
+            return redirect()->action('ServicesController@show', ['id' => $serviceId])->with('error', 'Jums nav tiesības izdzēst cita lietotāja atsauksmi');
+        }
         $feedback->delete();
-        return redirect()->action('ServicesController@show', ['id' => $serviceId])->with('success', 'Atsauksme dzēsta');
+        return redirect()->action('ServicesController@show', ['id' => $serviceId])->with('success', 'Atsauksmes dzēšana veiksmīgi izpildīta');
     }
 
+    // Pievienot favorītiem
     public function addToFavorites($id){
         $user = Auth::user();
+        // Pārliecināties, ka šis pakalpojums jau nav favorītos
         $favorite = Favorites::where("user_id", '=', $user->id)->where('service_id','=',$id)->get();
         if (count($favorite)) {
             return redirect()->action('ServicesController@show', ['id' => $id])->with('error', 'Pakalpojums jau ir favorītos');
         }
+        // Pievienot favorītiem
         $favorite = new Favorites;
         $favorite->user()->associate(User::find(Auth::user()->id));
         $favorite->service()->associate(Service::find($id));
         $favorite->save();
-        return redirect()->action('ServicesController@show', ['id' => $id])->with('success', 'Pakalpojums pievienots favorītiem');
+        return redirect()->action('ServicesController@show', ['id' => $id])->with('success', 'Pakalpojums veiksmīgi pievienots favorītiem');
     }
 
+    // Izņemt no favorītiem
     public function removeFromFavorites($id){
         $favorite = Favorites::find($id);
         $favorite->delete();
-        return redirect()->action('ServicesController@favorites')->with('success', 'Pakalpojums dzēsts no favorītiem');        
+        return redirect()->action('ServicesController@favorites')->with('success', 'Pakalpojums ir dzēsts no favorītiem');        
     }
 }
